integrate services with user dependency

+++ backend/app/services/report_service.py
@@ -18,13 +18,13 @@ def create_report(report_create, db, user_id):
     db.refresh(new_report)
     return new_report
 
-def get_reports(db):
-    return db.query(Report).all()
+def get_reports(db, user_id):
+    return db.query(Report).filter(Report.author_id == user_id).all()
 
-def get_report(report_id, db):
-    return db.query(Report).filter(Report.id == report_id).first()
+def get_report(report_id, db, user_id):
+    return db.query(Report).filter(Report.id == report_id, Report.author_id == user_id).first()
 
-def remove_report(report_id, db):
-    rows_deleted = db.query(Report).filter(Report.id == report_id).delete()
+def remove_report(report_id, db, user_id):
+    rows_deleted = db.query(Report).filter(Report.id == report_id, Report.author_id == user_id).delete()
     db.commit()
     return rows_deleted
